fix: Move all pages under [locale] to fix Vercel deployment
The root layout is now a pass-through for next-intl, so all pages must
live under src/app/[locale]/ to get the full layout with providers
(Clerk, NextIntl, Sentry, fonts). Without this, pages rendered without
any providers causing build/runtime failures.
