+++ src/app/[locale]/admin/AdminPageClient.tsx
