+++ src/app/[locale]/admin/hooks/useAdminStats.ts
