+++ src/app/[locale]/admin/page.tsx
