+++ src/app/[locale]/admin/views/ContactsView.tsx
