+++ src/app/[locale]/admin/views/DashboardView.tsx
