+++ src/app/[locale]/dashboard/[[...rest]]/page.tsx
