+++ src/app/[locale]/gdpr/page.tsx
