+++ src/app/[locale]/joaquin/page.tsx
