+++ src/app/[locale]/partidos/page.tsx
