+++ src/app/[locale]/sign-up/[[...rest]]/page.tsx
