+++ src/app/[locale]/trivia/page.tsx
