+++ tests/unit/app/joaquin.test.tsx
@@ -39,7 +39,7 @@ describe("Joaquín Page", () => {
   let renderResult: RenderResult;
 
   beforeEach(async () => {
-    const JoaquinPage = (await import("@/app/joaquin/page")).default;
+    const JoaquinPage = (await import("@/app/[locale]/joaquin/page")).default;
     renderResult = render(<JoaquinPage />);
   });
