+++ tests/unit/app/jugadores-historicos.test.tsx
@@ -39,7 +39,7 @@ vi.mock("lucide-react", () => ({
 describe("Jugadores Históricos Page", () => {
   beforeEach(async () => {
     const JugadoresHistoricos = (
-      await import("@/app/jugadores-historicos/page")
+      await import("@/app/[locale]/jugadores-historicos/page")
     ).default;
     render(<JugadoresHistoricos />);
   });
@@ -240,7 +240,7 @@ describe("Jugadores Históricos Page", () => {
     it("should use branded design patterns", async () => {
       // Re-render to get container reference
       const JugadoresHistoricos = (
-        await import("@/app/jugadores-historicos/page")
+        await import("@/app/[locale]/jugadores-historicos/page")
       ).default;
       const { container } = render(<JugadoresHistoricos />);
