+++ tests/unit/app/page.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen, within, waitFor } from "@testing-library/react";
 import { describe, it, expect, vi } from "vitest";
-import Home from "../../../src/app/page";
+import Home from "../../../src/app/[locale]/page";
 
 // Mock components and modules
 vi.mock("@/components/hero/HeroCommunity", () => ({
@@ -13,19 +13,29 @@ vi.mock("@/components/match/UpcomingMatchesWidget", () => ({
 vi.mock("@/components/widgets/ClassificationWidget", () => ({
   default: vi.fn(() => <div data-testid="mock-classification-widget" />),
 }));
+vi.mock("@/components/widgets/BetisEfemerides", () => ({
+  default: vi.fn(() => <div data-testid="mock-betis-efemerides" />),
+}));
 // Mock feature flags to return true (features enabled)
 vi.mock("@/lib/features/featureFlags", () => ({
   hasFeature: vi.fn(() => true),
 }));
 
+// Helper to render the async server component
+async function renderHome() {
+  const params = Promise.resolve({ locale: "es" });
+  const Component = await Home({ params });
+  return render(Component as React.ReactElement);
+}
+
 describe("Home page", () => {
-  it("renders HeroCommunity component", () => {
-    render(<Home />);
+  it("renders HeroCommunity component", async () => {
+    await renderHome();
     expect(screen.getByTestId("mock-hero-community")).toBeInTheDocument();
   });
 
   it("renders UpcomingMatchesWidget and ClassificationWidget when features are enabled", async () => {
-    render(<Home />);
+    await renderHome();
     // Wait for dynamic imports to load
     await waitFor(() => {
       expect(
@@ -39,8 +49,8 @@ describe("Home page", () => {
     });
   });
 
-  it('renders the "Join Us" section with correct text and links', () => {
-    render(<Home />);
+  it('renders the "Join Us" section with correct text and links', async () => {
+    await renderHome();
     // Text is split across lines with <br /> so we check for parts
     expect(screen.getByText(/¿Estás de visita/i)).toBeInTheDocument();
     expect(screen.getByText(/en Escocia\?/i)).toBeInTheDocument();
@@ -51,11 +61,10 @@ describe("Home page", () => {
       screen.getByText(/Todos los béticos son bienvenidos/i),
     ).toBeInTheDocument();
     expect(screen.getByRole("link", { name: /Únete/i })).toBeInTheDocument();
-    // Facebook and YouTube links are wrapped in FeatureWrapper for social media feature
   });
 
-  it("renders contact info section with correct details", () => {
-    render(<Home />);
+  it("renders contact info section with correct details", async () => {
+    await renderHome();
     // Find the "Ubicación" section and query within it
     const locationSection = screen.getByText("📍 Ubicación").closest("div");
     expect(locationSection).toBeInTheDocument();
